fixing variable declaration

+++ _exec.c
@@ -9,7 +9,7 @@
 int _exec(int  ac, char *av[])
 {
 	(void) ac;
-	size_t pid = 0;
+	int pid = 0;
 	int status = 0;
 	unsigned int i = 0;
 	struct stat sb;
